Separate server and app for future testing
